requirements.txt
sentence-transformers
flask
python-dotenv
requests
langchain
openai
pinecone-client
scikit-learn
numpy
pytest
pytest-flask
gunicorn
